Introduce PluginConverterFactory
